Factor connected devices into scheduling
Add two new sorting functions to the list of sorts to perform. One
of them sorts based on the number of connected devices falling
under a specified broad category. The other sorts based on the
number of specifically requested devices connected to the gateway.
The latter is higher in priority.

I am not entirely sure how this will shift the scheduling outcome.
If this does not perform well enough, then that means there are too
many criteria being simultaneously sorted for. If so, then we can
switch to a generic tiering approach wherein we create groups based
on criteria fulfillment and sort within those broad groups. Such an
approach will only become effective for larger deployments, though.